feat: fixing the validations on controllers

+++ app/Http/Controllers/ClientsController.php
@@ -13,30 +13,42 @@ class ClientsController extends Controller
     public function getAll(){
         try{
             $clients = Clients::all();
+            if(count($clients) === 0){
+                return response()->json(['error' => false, 'message' => 'There are no clients'], 200);
+            }
             return response()->json($clients, 200);
         }catch (QueryException $exception){
-            return response()->json(['error' => 'DATABASE ERROR'], 500);
+            return response()->json(['error' => true, 'message' => 'DATABASE ERROR'], 500);
         }
     }
+
     public function get($id){
         try{
             $client = Clients::find($id);
+            if(empty($client)){
+                return response()->json(['error' => true, 'message' => 'CLIENT NOT FOUND'], 400);
+            }
             return response()->json($client, 200);
         }catch (QueryException $exception){
-            return response()->json(['error' => 'CLIENT NOT FOUND'], 500);
+            return response()->json(['error' => true, 'message' => 'CLIENT NOT FOUND'], 500);
         }
     }
 
     public function store(Request $req){
-         $this->validate(
-            $req,
+        $validator = Validator::make(
+            $req->all(),
             [
                 'name' => 'required|max:40|min:3',
-                'email' => 'required|regex:/^.+@.+$/i',
+                'email' => 'required|email',
                 'phone' => 'required|max:13|min:10',
-                'cpf' => 'required|max:11|min:11'
+                'cpf' => 'required|string|cpf|unique:clients|max:11|min:11'
             ]
         );
+
+        if($validator->fails()){
+            return response()->json($validator->errors(), 400);
+        }
+
         try{
             $client = new Clients();
             $client->fill($req->all());
@@ -48,21 +60,24 @@ class ClientsController extends Controller
     }
 
    public function update($id, Request $req){
-            $this->validate(
-                $req,
+
+            $validator = Validator::make(
+                $req->all(),
                 [
                     'name' => 'max:40|min:3',
-                    'email' => 'regex:/^.+@.+$/i',
+                    'email' => 'email',
                     'phone' => 'max:13|min:10',
-                    'cpf' => 'max:11|min:11'
+                    'cpf' => 'string|unique:clients|cpf|max:11|min:11'
                 ]
             );
 
-            try{
+            if($validator->fails()){
+                return response()->json($validator->errors(), 400);
+            }
 
+            try{
                 $client = Clients::find($id)
                     ->update($req->all());
-
                 return response()->json(Clients::find($id), 200);
             }catch (QueryException $exception){
                 return response()->json(['error' => 'CLIENT NOT FOUND'], 500);
@@ -75,6 +90,9 @@ class ClientsController extends Controller
 
         try{
             $client = Clients::where('id', $id)->first();
+            if(empty($client)){
+                return response()->json(['error' => true, 'message' => 'CLIENT NOT FOUND'], 400);
+            }
             $client->delete();
             return response()->json(['message' => 'CLIENT DELETED'], 200);
         }catch (QueryException $exception){
